V1.2024.01.09: implementación del caso de uso Adquirir una Línea, que genera un contract y un customerLine. Otros cambios realizados.

+++ src/main/java/es/uca/iw/telefonuca/line/domain/Line.java
@@ -67,7 +67,7 @@ public class Line {
         this.description = description;
     }
 
-    public float getPricePerMinute() {
+    public int getPricePerMinute() {
         return pricePerMinute;
     }
 
@@ -75,7 +75,7 @@ public class Line {
         this.pricePerMinute = pricePerMinute;
     }
 
-    public float getPricePerMegabyte() {
+    public int getPricePerMegabyte() {
         return pricePerMegabyte;
     }
